+++ resources/assets/js/components/enso/bulma/Overlay.vue
@@ -13,6 +13,7 @@
 <script>
 
 export default {
+    name: 'Overlay',
 
     props: {
         size: {