+++ resources/assets/js/pages/App.vue
@@ -1,13 +1,12 @@
 <template>
 
     <div id="app">
-        <transition
-            enter-active-class="fadeIn"
+        <transition enter-active-class="fadeIn"
             leave-active-class="fadeOut">
             <component :is="component"
                 class="animated"
                 @enter-app="showHome = false"
-                @login="showHome=true;setState()">
+                @login="showHome = true; setState()">
             </component>
         </transition>
     </div>