+++ resources/assets/js/pages/layout/VueAside.vue
@@ -1,6 +1,6 @@
 <template>
 
-    <div>
+    <div class="aside">
         <slot></slot>
     </div>
 
@@ -14,27 +14,27 @@ export default {
 
 </script>
 
-<style>
+<style lang="scss">
 
-    .aside {
+    div.aside {
+        p.menu-label {
+            font-size: 0.8rem;
+        }
+
         position: fixed;
         bottom: 0;
         top: 50px;
-        padding-top: 25px;
-        padding-bottom: 50px;
+        padding-top: 20px;
+        padding-bottom: 20px;
         width: 180px;
         -webkit-box-shadow: 0px 0px 5px 1px rgba(133,133,133,1);
         -moz-box-shadow: 0px 0px 5px 1px rgba(133,133,133,1);
         box-shadow: 0px 0px 5px 1px rgba(133,133,133,1);
-        font-size: 0.9rem;
+        font-size: 0.9em;
         max-height: 100vh;
-        z-index: 1025;
+        z-index: 2;
         overflow-y: auto;
         overflow-x: hidden;
-    }
-
-    .aside p.menu-label {
-        font-size: 0.8rem;
     }
 
 </style>