+++ resources/assets/js/pages/layout/navbar/SettingsControl.vue
@@ -1,0 +1,9 @@
+<template>
+    <a class="navbar-item"
+        @click="$store.commit('layout/settingsBar/toggle')">
+        <span class="icon is-small">
+            <i class="fa fa-gears">
+            </i>
+        </span>
+    </a>
+</template>