+++ resources/assets/js/pages/layout/settings/ThemeSelector.vue
@@ -1,6 +1,6 @@
 <template>
 
-    <div class="level">
+    <div class="level is-mobile settings">
         <div class="level-left">
             <div class="level-item">
                 {{ title }}