+++ routes/api.php
@@ -6,10 +6,14 @@ Route::get('/getAppName', function () {
 
 Route::namespace('Auth')
     ->group(function () {
-        Route::post('login', 'LoginController@login');
-        Route::post('logout', 'LoginController@logout')->name('logout');
-        Route::post('password/email', 'ForgotPasswordController@sendResetLinkEmail')->name('password.email');
-        Route::post('password/reset', 'ResetPasswordController@reset');
+        Route::post('login', 'LoginController@login')
+            ->name('login');
+        Route::post('logout', 'LoginController@logout')
+            ->name('logout');
+        Route::post('password/email', 'ForgotPasswordController@sendResetLinkEmail')
+            ->name('password.email');
+        Route::post('password/reset', 'ResetPasswordController@reset')
+            ->name('password.reset');
     });
 
 Route::middleware(['auth'])